미니프로젝트_3차/code_zip/pyproject.toml
[project]
name = "miniproj3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "faiss-cpu>=1.12.0",
    "google-adk>=1.12.0",
    "google-genai>=1.31.0",
    "ipykernel>=6.30.1",
    "litellm>=1.76.0",
    "openai>=1.101.0",
    "pypdf>=6.1.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "yfinance>=0.2.66",
]
